add meat/user info page for admin complete
